getCounts generalized
